web_app/requirements.txt
asgiref==3.6.0
astroid==2.13.2
autopep8==2.0.1
dill==0.3.6
Django==4.1.5
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==20.1.0
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.4.0
platformdirs==2.6.2
pycodestyle==2.10.0
pylint==2.15.10
pytz==2022.7.1
SHop==0.0.2
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
